requirements.txt
pillow
opencv-python
google-cloud-vision
python-Levenshtein